% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_node_id.R
\name{get_station_node_id}
\alias{get_station_node_id}
\title{get_station_node_id}
\usage{
get_station_node_id(station, grph)
}
\arguments{
\item{grph}{}
}
\description{
Internal function to get the node id from the lu_graph by station name
}
