% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tuble.R
\name{guess_tuble}
\alias{guess_tuble}
\title{guess_tuble}
\usage{
guess_tuble(station, n_stops, prev_set, grph)
}
\arguments{
\item{grph}{}
}
\description{
Given a tuble guess, i.e. a station and the number of stops, n_stops, to the
answer/target station, will return a set of stations that satisfy those two criteria.
}
\details{
If a set of stations is provided, i.e. the result of a previous guess, the
intersection of that set with the results of this guess will be returned
}
